requirements.txt
django-notification==1.0
django-voting==0.2
